% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{LAU2_ccaa_2020_canarias}
\alias{LAU2_ccaa_2020_canarias}
\title{Geometrías para las CC.AA, 2020, de LAU2boundaries4Spain. (Spain)}
\format{
An sf data.frame with 19 rows and 8 variables
\itemize{
\item ine_ccaa: Código INE de la CA (2 dígitos). Character
\item ine_ccaa.n: Nombre de la CA. Character
\item pob_2020: Población por CC.AA  total. Numeric
\item ine_ccaa.n.pjp: Nombre de la CA no oficial pero good. Character
\item X: Centroide (X). Numeric
\item Y: Centroide (Y). Numeric
\item geometry: geometría provincial. sfc_GEOMETRY
\item X1: Centroide (X) pero desplazado para Ceuta y Melilla. Numeric
\item Y1: Centroide (Y) pero desplazado para Ceuta y Melilla. Numeric
}
}
\source{
\url{https://github.com/rOpenSpain/LAU2boundaries4spain}
}
\usage{
LAU2_ccaa_2020_canarias
}
\description{
Geometrías municipales 2020 (SPAIN). Sacados del package LAU2boundaries4Spain.
Recuerda que he desplazado Canarias
Recuerda que hay dos tipos de centroides.
X e Y (sin desplazar, bueno Canarias sí está desplazada).
X1 e Y1 (desplazados Ceuta y Melilla para q se vean bien e los mapas)
}
\examples{
\dontrun{
 LAU2_ccaa_2020_canarias <- LAU2_ccaa_2020_canarias
}

}
\keyword{datasets}
