% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjpv.datos.01-package.R
\docType{package}
\name{pjpv.datos.01-package}
\alias{pjpv.datos.01}
\alias{pjpv.datos.01-package}
\title{pjpv.datos.01: Paquete de uso personal}
\description{
Paquete para uso personal con conjuntos de datos
}
\author{
\strong{Maintainer}: Pedro Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
